+++ RELEASES.md
@@ -3,6 +3,7 @@
 ## Version 2.0
 ### TODO
 * (DONE) Fix utility memoryfile examples
+* (PENDING) pushCall instead of pushReinit
 * (PENDING) Test serialization for all supported std containers
 * (PENDING) Finalizing the IPC library - what is planned for SFv2.0.
 * Cross-platform support: